Modify MaterialRealAux
* Changed the 'matpro' parameter in MaterialRealAux to 'property'.
* Put functionality of MaterialRealScaledAux (ELK) into MaterialRealAux
* Deleted MaterialRealScaledAux

All tests pass except those in Pronghorn.  All tests fail there due to a
registration problem (not my fault).

I recommend calling setCheckUnusedFlag(true); in your application.  This
causes an error to be raised if the input file contains a line the code
does not recognize (like 'matpro = ...' now that it should be 'property
= ...'.

Closes #1204

r11788